+++ src/main/java/io/github/zeroone3010/yahueapi/DaylightSensorImpl.java
@@ -5,8 +5,12 @@ import io.github.zeroone3010.yahueapi.domain.SensorDto;
 import java.net.URL;
 import java.util.Map;
 import java.util.function.Supplier;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 final class DaylightSensorImpl extends BasicSensor implements DaylightSensor {
+  private static final Logger logger = Logger.getLogger("DaylightSensor");
+
   DaylightSensorImpl(final String id, final SensorDto sensor, final URL url, final Supplier<Map<String, Object>> stateProvider) {
     super(id, sensor, url, stateProvider);
   }
@@ -22,6 +26,11 @@ final class DaylightSensorImpl extends BasicSensor implements DaylightSensor {
 
   @Override
   public boolean isDaylightTime() {
-    return readStateValue("daylight", Boolean.class);
+    try {
+      return readStateValue("daylight", Boolean.class);
+    } catch (final NullPointerException npe) {
+      logger.log(Level.WARNING, "It appears that the daylight sensor has not been configured.");
+      return false;
+    }
   }
 }